Tile elevators adjust their wall heights now

+++ Dungeoneer/src/com/interrupt/dungeoneer/tiles/TileSurface.java
@@ -5,8 +5,8 @@ public class TileSurface {
     // TODO: Someday move the texture and texture atlas into here as well
 
     // Texture Offsets
-    public byte yOffset = 0;
-    public byte xOffset = 0;
+    public float yOffset = 0;
+    public float xOffset = 0;
 
     public TileSurface() { }
 }
